requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
tabulate>=0.9.0
python-dotenv>=1.0.0
